Refactor a lot of code
